add delete country

+++ utils/deleteCity.js
@@ -1,9 +1,23 @@
 import { removeImage } from './removeImage';
 
 export const deleteCity = async (title, supabase) => {
-  await removeImage(`public/${title}/image`, supabase, 'itineraries');
+  const { data: itinerary } = await supabase
+    .from('itineraries')
+    .select('title')
+    .eq('city', title)
+    .single();
 
-  await supabase.from('itineraries').delete().eq('city', title);
+  if (itinerary) {
+    await removeImage(
+      `public/${itinerary.title}/image`,
+      supabase,
+      'itineraries',
+    );
+
+    await supabase.from('itineraries').delete().eq('city', title);
+  }
+
+  await removeImage(`public/${title}/image`, supabase, 'cities');
 
   const { data, error } = await supabase
     .from('cities')
